web/tickets-please/requirements.txt
pyjwt==2.4.0
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
